requirements.txt
opencv-python
kafka-python
boto3
python-dotenv
pillow